summarization_pipeline_app/requirements.txt
numpy
pandas
torch
tf-keras
flask
transformers
flask-cors
python-dotenv
-e .
